refactor: move storage into a new crate
