+++ crates/storage/src/state.rs
@@ -1316,7 +1316,7 @@ impl CanonicalBlocksTable {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::storage::Storage;
+    use crate::Storage;
 
     mod contracts {
         use super::*;
@@ -1612,7 +1612,7 @@ mod tests {
 
     mod starknet_blocks {
         use super::*;
-        use crate::storage::test_utils;
+        use crate::test_utils;
 
         fn create_blocks() -> [StarknetBlock; test_utils::NUM_BLOCKS] {
             test_utils::create_blocks()
@@ -2034,7 +2034,7 @@ mod tests {
 
     mod starknet_events {
         use super::*;
-        use crate::storage::test_utils;
+        use crate::test_utils;
         use pathfinder_common::starkhash;
         use pathfinder_common::{EntryPoint, EventData, Fee};
         use web3::types::H128;
@@ -2647,7 +2647,7 @@ mod tests {
 
     mod starknet_updates {
         use super::*;
-        use crate::storage::fixtures::with_n_state_updates;
+        use crate::fixtures::with_n_state_updates;
 
         mod get {
             use super::*;
